# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Galaktika galaxy game",
        "title": "Galaktika API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/fleet-builds": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet-builds"
                ],
                "summary": "List all fleet builds",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/galaxy.FleetBuild"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet-builds"
                ],
                "summary": "Create a fleet build",
                "parameters": [
                    {
                        "description": "FleetBuild data",
                        "name": "fleetBuild",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/galaxy.FleetBuild"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/galaxy.FleetBuild"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/fleet-builds/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet-builds"
                ],
                "summary": "Get a fleet build by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FleetBuild ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/galaxy.FleetBuild"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet-builds"
                ],
                "summary": "Update a fleet build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FleetBuild ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "FleetBuild data",
                        "name": "fleetBuild",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/galaxy.FleetBuild"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/galaxy.FleetBuild"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet-builds"
                ],
                "summary": "Delete a fleet build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FleetBuild ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/fleet-builds/{id}/ship-models": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet-builds"
                ],
                "summary": "List ship models assigned to a fleet build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FleetBuild ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/galaxy.FleetBuildToShipModel"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet-builds"
                ],
                "summary": "Assign a ship model to a fleet build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FleetBuild ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Assignment data",
                        "name": "assignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/galaxy.FleetBuildToShipModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/galaxy.FleetBuildToShipModel"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/galaxy.FleetBuildToShipModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/fleet-builds/{id}/ship-models/{shipModelId}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet-builds"
                ],
                "summary": "Unassign a ship model from a fleet build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FleetBuild ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ShipModel ID",
                        "name": "shipModelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "galaxy.FleetBuild": {
            "type": "object",
            "properties": {
                "assignedShipModels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/galaxy.ShipModelAssignment"
                    }
                },
                "attack_resources": {
                    "type": "number"
                },
                "cargo_resources": {
                    "type": "number"
                },
                "defense_resources": {
                    "type": "number"
                },
                "division_id": {
                    "type": "string"
                },
                "engine_resources": {
                    "type": "number"
                },
                "id": {
                    "description": "stored to DB",
                    "type": "string"
                },
                "race_id": {
                    "type": "string"
                },
                "usedResources": {
                    "type": "number",
                    "format": "float64"
                }
            }
        },
        "galaxy.FleetBuildToShipModel": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "fleet_build_id": {
                    "type": "string"
                },
                "result_mass": {
                    "type": "number"
                },
                "shipModel": {
                    "description": "not stored to DB directly",
                    "allOf": [
                        {
                            "$ref": "#/definitions/galaxy.ShipModel"
                        }
                    ]
                },
                "ship_model_id": {
                    "type": "string"
                }
            }
        },
...
        "galaxy.ShipModel": {
            "type": "object",
            "properties": {
                "cargo_mass": {
                    "type": "number"
                },
                "defense_mass": {
                    "type": "number"
                },
                "engine_mass": {
                    "type": "number"
                },
                "guns": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "one_gun_mass": {
                    "type": "number"
                },
                "owner_id": {
                    "type": "string"
                }
            }
        },
        "galaxy.ShipModelAssignment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "shipModel": {
                    "$ref": "#/definitions/galaxy.ShipModel"
                }
            }
        },
...
    }
}